Convert to `record_ml_event` method for sending AI events (#31)
* Ported Bedrock POC code to SDK+Agent.
Building example file for bedrock that runs with the NR Agent
and the NR SDK for LLMs.

Currently testing only with Amazon Titan LLM.

* Added support for Anthropic Clause with three different models.

* Adding support for AI21 Jurassic models.

* Added support for Cohere Command model.

* Ensure only one set of model logic occurs each run.
Moved several attributes to have a guard check so that
we do not send `None` values.

* Fix example script so that each run of Bedrock occurs in a separately
monitored code block from the New Relic Agent. This stops all of the messages
from appearing for each request.

* Adding in the default max tokens used for each model. This value
can be overridden on each call.

* Adding a Transaction Begin event that is required with the new Responses UI.
Switching the first message for Bedrock from 'system' to 'user'.

* Adding logic to account for a few missing variables. TransactionId needs to be on
each message. Is final message should be set only on the last message. Fixed the
adding of tokens so it only happens if both values are set.

* Adding additional function annotation as recommended by Python Agent team.
Creating a sample app the run OpenAI and Bedrock with the same requests to
allow comparison.

* Trying to get the APM name right.

* Noam/streaming (#26)

* refactor:extract code to multiple files

* monitor stream for chat completion

* rename file

* async chat completion

* fix stream async + add spans for streams

* remove unused code

* Added stream token calculation

* fix typo + remove prints and comments

* added "stream" attribute to completion events

* handle errors before the stream starts

* Making sure the NEW_RELIC_APP_NAME env is set. Passing it as a parameter
to the agent didn't seem to work well.

* Updating LangChain example to work with OpenAI and Bedrock.

* Creating events from LangChain.

* Removed log statements from the bedrock code to clean up STDOUT.
Added additional model parsing for langchain when using Bedrock LLM.

* Updating langchain example to use BedrockChat.

* Adding in token calculations for both Anthropic and AI21 models.

* Added very high level error event for when an exception is raised.

* Adding custom metrics to track library versions. This should/will be used
later to set an attribute to know that the APM entity is sending AI metrics.

* Fixing the name of the pinecone client.

* Removing a now dead function. Minor tweak to add timestamp to an event.

* Disabled the module search path modification by default as it is not needed for
users of the repo. Only developers trying to test changes need it enabled.

* Removing unused import.

* Moving to another API to submit events so that values can be larger
than 255 characters.

* Moving a log line to debug so it does not come up for users.

* Making bedrock dependencies as optional. Putting a check in the code
to ensure minimum versions of botocore and boto3 before trying to
instrument Bedrock APIs.

* Trying new version of newrelic agent.

* Move anthropic imports to where it is needed so that everyone
does not have to install it.

* Applying feedback from PR review. Bumping up the minimum version of
LangChain so that it can actually function with the BedrockChat wrapper.

* Removing test code for early development.
Fixed a merge conflict that dropped a few lines of code.

* Adding a missing argument to a method that will allow tracking completions correctly.

* Merged latest from staging.
Moved to the proper New Relic Python Agent branch.
Applied consistent trace naming for Bedrock messages which mirrors OpenAI.

* fix: Dependency

---------

Co-authored-by: nhoffmann-newrelic <123460916+nhoffmann-newrelic@users.noreply.github.com>
Co-authored-by: Ryan Kadri <rkadri@newrelic.com>